custom-elements/instantiating-custom-elements/extensions-to-document-interface/create-element-is-attribute.html

<!DOCTYPE html>
<html>
<head>
<title>createElement should set IS attribute to type</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If TYPE is not the same as localName, set the value of ELEMENT's is attribute to TYPE">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    HTML5_ELEMENTS.forEach(function (value) {        
        var obj = doc.createElement(value);
        var name = 'x-1-' + value;        
        var proto = Object.create(obj.constructor.prototype);
        doc.register(name, {prototype: proto, extends: value});
        var customElement = doc.createElement(value, name);
            
        assert_equals(customElement.getAttribute('is'), name, 
                'Wrong value of the IS attribute');            
    });    
}, 'If type is not the same as localName, the value of ELEMENT\'s is attribute ' +
    'must be set to type by createElement');


test(function() {
    var doc = newHTMLDocument();
    
    HTML5_ELEMENTS.forEach(function (value) {        
        var obj = doc.createElement(value);
        var name = 'x-2-' + value;        
        var proto = Object.create(obj.constructor.prototype);
        doc.register(name, {prototype: proto, extends: value});
        var customElement = doc.createElementNS(HTML_NAMESPACE, value, name);
            
        assert_equals(customElement.getAttribute('is'), name, 
                'Wrong value of the IS attribute');            
    });        
}, 'If type is not the same as localName, the value of ELEMENT\'s is attribute ' +
    'must be set to type by createElementNS');
    
</script>
</body>
</html>
